pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain_bee"
version = "1.0.0"
description = "Multi-Agent AI Framework — Architecture Demo"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "KimiinCHN-AGI"}
]
keywords = ["agent", "multi-agent", "ai", "architecture-demo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
brain_bee = "brain_bee.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["brain_bee*"]

[tool.setuptools.package-data]
brain_bee = ["manifests/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
